Fixing the FileEntry require statement.
For some reason, as soon as the objective-c was fixed, it exposed an issue whereby cordovaHTTP was requiring its dependancy ('cordova-plugin-file’) incorrectly. After looking into this, it seems as though cordovaHTTP might be referncing an older version of this plugin. If you look cordovaHTTP’s plugin.xml it says it depends upon commit ‘r0.2.5’ of ‘org.apache.cordova.file’ where as we are using version 2.0.0 of 'cordova-plugin-file’. (Note the change of plugin name - I believe the plugin naming style changed when Cordova switched over to using NPM to manage their packages.)